refactor : view와 controller 분리
