+++ src/js/index.js
@@ -1,9 +1,7 @@
 import Calculator from "./controller.js";
-import CalculatorView from "./view.js";
 
 const init = () => {
-  const calculator = new Calculator();
-  new CalculatorView(calculator);
+  new Calculator();
 };
 
 init();
